requirements.txt
Django==1.6.5
Unipath==1.0
amqp==1.4.6
anyjson==0.3.3
argparse==1.2.1
billiard==3.3.0.18
boto==2.32.1
celery==3.1.7
dj-database-url==0.3.0
dj-static==0.0.6
django-bootstrap3==4.11.0
django-celery==3.1.16
django-registration==1.0
django-ses==0.6.0
django-storages==1.1.8
django-toolbelt==0.0.1
gspread==0.2.2
gunicorn==19.1.1
kombu==3.0.23
nltk==3.0.0
numpy==1.9.0
pandas==0.15.2
psycopg2==2.5.4
pusher==0.8
pusher-rest==0.1.0
python-dateutil==2.4.0
python-postmark==0.4.6
pytz==2014.7
redis==2.10.3
six==1.9.0
static3==0.5.1
wsgiref==0.1.2
